fix: flip dragFeedbackOffset.y when piece is flipped
Noticed this while building the "over the board" screen for lichess
mobile v2. When pieces are flipped, the person dragging them will
probably look at the app in such a way, that the piece is *not* flipped
from their POV. This means with the previous behavior, the drag feedback
offset would be shifted into the wrong vertical direction.